requirements.txt
python-bale-bot==2.5.0
requests
Flask
